feat: reorganized the protocol into different modules
