+++ acapy_plugin_pickup/__init__.py
@@ -2,13 +2,9 @@
 
 from aries_cloudagent.config.injection_context import InjectionContext
 from aries_cloudagent.core.protocol_registry import ProtocolRegistry
-from .protocol import (
-    LiveDeliveryChange,
-    MessagesReceived,
-    Status,
-    StatusRequest,
-    DeliveryRequest,
-)
+from .protocol.live_mode import LiveDeliveryChange
+from .protocol.delivery import DeliveryRequest, Delivery, MessagesReceived
+from .protocol.status import StatusRequest, Status
 
 
 async def setup(context: InjectionContext):
@@ -19,6 +15,7 @@ async def setup(context: InjectionContext):
         {
             Status.message_type: Status,
             StatusRequest.message_type: StatusRequest,
+            Delivery.message_type: Delivery,
             DeliveryRequest.message_type: DeliveryRequest,
             MessagesReceived.message_type: MessagesReceived,
             LiveDeliveryChange.message_type: LiveDeliveryChange,
